% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{add_cartesian_plane}
\alias{add_cartesian_plane}
\title{Add cartesian system to a plot}
\usage{
add_cartesian_plane(at.y = 0, at.x = 0, col = "black", lty = "solid",
  lwd = par("lwd"))
}
\arguments{
\item{at.y}{y centre of the system}

\item{at.x}{x centre of the system}

\item{col}{grid color}

\item{lty}{line type}

\item{lwd}{line width}
}
\value{
The function adds a cartesian plane to a plot using
\code{\link{abline}}.
}
\description{
Add a cartesian system to a plot using \code{\link{abline}}.
}

