% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{add_to_wb}
\alias{add_to_wb}
\title{Faster export with openxlsx}
\usage{
add_to_wb(wb = NULL, sheet = NULL, x = NULL, ...)
}
\arguments{
\item{wb}{a Workbook object}

\item{sheet}{name of the sheet to be created}

\item{x}{object to be exported}

\item{...}{further arguments passed to writeData}
}
\description{
This is a wrapper around addWorksheet and writeData to allow
less typing. It export one object per sheet, the sheet will be
created accordingly to the name given
}

