% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{altro_specificare}
\alias{altro_specificare}
\title{Handle other specify questions in a sensible way}
\usage{
altro_specificare(x)
}
\arguments{
\item{x}{a character vector}
}
\description{
Handle other specify questions in a sensible way
}
