% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{as.numeric2}
\alias{as.numeric2}
\title{coerce a character numeric with , to numeric}
\usage{
as.numeric2(x)
}
\arguments{
\item{x}{character numeric with , to numeric}
}
\description{
coerce a character numeric with , to numeric
}
