% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicero.R
\name{cicero}
\alias{cicero}
\title{cicero: some similarities with moustache's idea but designed to
be LaTeX friendly and tiny}
\usage{
cicero(tmpl = NULL, strict = FALSE, ...)
}
\arguments{
\item{tmpl}{template with \code{<<>>} as delimiters}

\item{strict}{data provided must have all the same length
(otherwise recycling will act)}

\item{...}{data (vectors) passed with the same name as the variable
to be filled in tmpl. If missing lookup is made in the calling
environment}
}
\description{
cicero: some similarities with moustache's idea but designed to
be LaTeX friendly and tiny
}
\examples{

tmp <- "Hi my name's  <<nome>> \\textbf{<<cognome>>}.
Born on <<dat_nas>>. Heigth <<heigth>>, weight <<weight>>."
cicero(tmpl = tmp,
       nome = "mario",
       heigth = 1.78,
       cognome = "rossi",
       dat_nas = '2013-10-10',
       weight = 80)

nome <- "mario"
cognome <- c("rossi", "bianchi")
heigth <- 1.78
dat_nas <- '2013-10-10'
weight <- 80
cicero(tmp, strict = FALSE)
\dontrun{cicero(tmp, strict = TRUE)}

}
