% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicero2.R
\name{cicero2}
\alias{cicero2}
\title{cicero: some similarities with moustache's idea but designed to
be LaTeX friendly and tiny}
\usage{
cicero2(x = NULL, strict = FALSE, ...)
}
\arguments{
\item{x}{template with \code{<<>>} as delimiters, or a function}

\item{strict}{data provided must have all the same length
(otherwise recycling will act)}

\item{...}{data (vectors) passed with the same name as the variable
to be filled in tmpl. If missing lookup is made in the calling
environment}
}
\description{
cicero: some similarities with moustache's idea but designed to
be LaTeX friendly and tiny
}
\examples{

## Example with character template
tmp <- "Hi my name's  <<nome>> \\textbf{<<cognome>>}.
Born on <<dat_nas>>. Heigth <<heigth>>, weight <<weight>>."
cicero2(x = tmp,
        nome = "mario",
        heigth = 1.78,
        cognome = "rossi",
        dat_nas = '2013-10-10',
        weight = 80)

nome <- "mario"
cognome <- c("rossi", "bianchi")
heigth <- 1.78
dat_nas <- '2013-10-10'
weight <- 80
cicero2(tmp, strict = FALSE)
\dontrun{cicero2(tmp, strict = TRUE)}

## Example with function as x
## a possible application for function argument: choose the
## template in runtime and fill it with proper data.
select <- function(nome, cognome, height, weight)
{
    tmpl1 <- "Hi, I'm  <<nome>> <<cognome>>. My bmi is ok: <<bmi>>."
    tmpl2 <- "Hi, I'm  <<nome>> <<cognome>>. My bmi is not ok: <<bmi>>."
    bmi <- weight/(height^2)
    if (bmi < 25) cicero2(x = tmpl1, nome = nome, cognome = cognome, bmi = bmi)
    else cicero2(x = tmpl2, nome = nome, cognome = cognome, bmi = bmi)
}
    
cicero2(x = select,
        nome = "mario", cognome = "rossi",
        height = 1.78, weight = 82)

cicero2(x = select,
        nome = "mario", cognome = "rossi",
        height = 1.78, weight = 75)
 
}
