% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{comma_char_to_numeric}
\alias{comma_char_to_numeric}
\title{coerce a character numeric with , to numeric}
\usage{
comma_char_to_numeric(x)
}
\arguments{
\item{x}{character numeric with , to numeric}
}
\description{
coerce a character numeric with , to numeric
}
