% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{comment_df}
\alias{comment_df}
\title{Comment several variables of a data.frame}
\usage{
comment_df(x, var_com)
}
\arguments{
\item{x}{a data.frame}

\item{var_com}{variable name/comment pair}
}
\description{
Comment several variables of a data.frame using a for loop of \code{comment}
}
\examples{
db <- data.frame(a = letters, b = LETTERS)
db <- comment_db(x = db, c('a', 'lowercase letters',
                           'b', 'uppercase letters'))
}

