% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose}
\alias{compose}
\title{compose functions passed to ...}
\usage{
compose(...)
}
\arguments{
\item{...}{functions to be composed}
}
\description{
compose functions passed to ...
}
\examples{
add2 <- function(x) x+2
mul2 <- function(x) x*2
f = compose(add2, mul2)
g = compose(mul2, add2)
f(1:3)
g(1:3)
}
