% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_utils.R
\name{date_mdy}
\alias{date_mdy}
\title{Create a \code{\link{Date}} from month, day and year.}
\usage{
date_mdy(month, day, year)
}
\arguments{
\item{month}{Numeric vector of months}

\item{day}{Numeric vector of days}

\item{year}{Numeric vector of years}
}
\value{
A vector of \code{\link{Date}}.
}
\description{
Create a \code{\link{Date}} from month, day and year.
}
\examples{
db <- data.frame(month = 10:12, day = 15:17, year = 1982:1984)
date_mdy(month = db$month, day= db$day, year=db$year)
}
