% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_utils.R
\name{date_plus}
\alias{date_plus}
\title{Add a period of time to a Date object}
\usage{
date_plus(x, years = 0, months = 0, days = 0)
}
\arguments{
\item{x}{starting Date}

\item{years}{years to be added}

\item{months}{months to be added}

\item{days}{days to be added}
}
\description{
Add a period of time to a Date object
}
\examples{
(today <- Sys.Date())
date_plus(today, years = 1, months = 1, days = 1)
}
