% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_utils.R
\name{datetime_import}
\alias{datetime_import}
\title{Quick alias for importing italian csv common datetime}
\usage{
datetime_import(x, fmt = "\%d/\%m/\%Y \%H.\%M.\%S", output = c("Date",
  "POSIXlt", "POSIXct"), ...)
}
\arguments{
\item{x}{character with date}

\item{fmt}{format given to as.Date}

\item{output}{class name of the ouput}

\item{...}{other arguments passed to as.POSIXlt}
}
\description{
Quick alias for importing italian csv common datetime
}
