% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_utils.R
\name{day}
\alias{day}
\title{Extract numeric days of month from a \code{Date},
\code{POSIXct} or \code{POSIXlt} vector.}
\usage{
day(x)
}
\arguments{
\item{x}{Date, \code{\link{POSIXct}} or \code{\link{POSIXlt}}
vector}
}
\value{
A numeric vector.
}
\description{
Extract numeric days of month from a \code{Date},
\code{\link{POSIXct}} or \code{\link{POSIXlt}} vector.
}
\examples{

Sys.Date()
day(Sys.Date())
day(as.POSIXct(Sys.Date()))
day(as.POSIXlt(Sys.Date()))


}
