% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_utils.R
\name{dmy_import}
\alias{dmy_import}
\title{Quick alias for importing italian csv common date formats}
\usage{
dmy_import(x, fmt = "\%d/\%m/\%Y", ...)
}
\arguments{
\item{x}{character with date}

\item{fmt}{format given to as.Date}

\item{...}{other arguments passed to as.Date}
}
\description{
Quick alias for importing italian csv common date formats
}
