% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummify.R
\name{dummify}
\alias{dummify}
\title{Generate a data.frame of dummy variables given a vector}
\usage{
dummify(x, prefix = "")
}
\arguments{
\item{x}{a vector}

\item{prefix}{name prefix of returned \code{\link{data.frame}} columns}
}
\value{
a \code{\link{data.frame}} to be used with \code{\link{cbind}}
}
\description{
Generate a data.frame of dummy variables given a vector
}
\examples{

rep(1:3, 4)
dummify(rep(1:3, 4))

}
