% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{duplicated2}
\alias{duplicated2}
\title{Determine all duplicate elements}
\usage{
duplicated2(x, all = TRUE, ...)
}
\arguments{
\item{x}{Numeric vector of centimeter.}

\item{all}{Logical: if TRUE considers duplicated each observation
present more than one time in the object, otherwise implement
base::duplicated algorithm (considers duplicated the second, third and so on
duplicated observation).}

\item{...}{Other arguments passed to base::duplicated.}
}
\value{
A logical vector with duplicated marked as TRUE.
}
\description{
base::duplicated determines which elements of a vector or data frame are
duplicates of elements with smaller subscripts, and returns a logical vector
indicating which elements (rows) are duplicates.
}
\details{
This wrapper allows to find all duplicated obs.
}
\examples{
duplicated( c(1,0, 2, 0, 3, 2)) # from base R
duplicated2(c(1,0, 2, 0, 3, 2)) # from lbmisc
}

