% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_if.R
\name{end_show_if}
\alias{end_show_if}
\title{end showing a chunk of a report if a logical condition is met}
\usage{
end_show_if(x)
}
\arguments{
\item{x}{a logical value}
}
\description{
end showing a chunk of a report if a logical condition is met; to
be used in a results = 'asis' chunk of Sweave/knitr
}
\examples{

eval_old_stuff <- FALSE
end_show_if(eval_old_stuff)

}
