% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eng_maxima.R
\name{eng_maxima}
\alias{eng_maxima}
\title{A maxima engine for knitr}
\usage{
eng_maxima(options)
}
\arguments{
\item{options}{knitr options}
}
\description{
A maxima engine for knitr
}
\examples{
\dontrun{
## Add this to your document
library(knitr)
knit_engines$set(maxima = yapomif::eng_maxima)
opts_chunk$set("engine" = "maxima")
opts_chunk$set("echo" = FALSE)
}
}
