% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment_period.R
\name{enrollment_period}
\alias{enrollment_period}
\title{Find the start and end of enrollment}
\usage{
enrollment_period(x, lang = c("en", "it"))
}
\arguments{
\item{x}{a Date of patient enrollment}

\item{lang}{output language}
}
\description{
Find the start and end of enrollment and report in month year char format
}
