% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{equal}
\alias{equal}
\title{test equality of two vectors without leaving NA}
\usage{
equal(x, y, one_NA = FALSE, both_NA = TRUE)
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}

\item{one_NA}{what to output if only one of the two provided is NA
(default = TRUE, aka value are diverse)}

\item{both_NA}{what to output if only one of the two provided is NA
(default = TRUE, aka value are diverse)}
}
\description{
test equality of two vectors without leaving NA as output (aka
considering NA = NA are equal, so TRUE, and value = NA are diverse,
so FALSE)
}
\examples{
a <- c(NA, NA, 1, 2)
b <- c(1,  NA, 2, 2)
equal(a, b)

}
