% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.grid2.R
\name{expand.grid2}
\alias{expand.grid2}
\title{A sensible expand.grid}
\usage{
expand.grid2(...)
}
\arguments{
\item{...}{params given to}
}
\description{
An expand.grid with better looking ordering and some defaults
(to date, stringsAsFactors = FALSE)
}
