% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{exporter}
\alias{exporter}
\title{Export a list of dataset to an xlsx file}
\usage{
exporter(x, f, ...)
}
\arguments{
\item{x}{openxlsx::write.xlsx x parameter}

\item{f}{openxlsx::write.xlsx file parameter}

\item{...}{further arguments passed to openxlsx::write.xlsx}
}
\description{
Export a list of dataset to an xlsx file
}
