% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{factor_blankNA}
\alias{factor_blankNA}
\title{Make a factor from character removing all blanks "" values (setting to NA)}
\usage{
factor_blankNA(x)
}
\arguments{
\item{x}{character}
}
\description{
Make a factor from character removing all blanks "" values (setting to NA)
}
