% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythontex_utils.R
\name{fig_dump}
\alias{fig_dump}
\title{save a figure done in R to a file with multiple formats}
\usage{
fig_dump(code, outfile = "outputs/test")
}
\arguments{
\item{code}{code to be evaluated (as string) to produce the figure}

\item{outfile}{path without extension}
}
\description{
save a figure done in R to a file with multiple formats
}
\examples{

fig_dump("roc_with_boxplot(sani_mci$score, sani_mci$sani_mci)", outfile = outfile)

}
