% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten}
\alias{flatten}
\title{flatten a list of nested list}
\source{
https://stackoverflow.com/questions/16300344
}
\usage{
flatten(x)
}
\arguments{
\item{x}{a list of nested list}
}
\description{
flatten a list of nested list
}
