% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{gdocs_download}
\alias{gdocs_download}
\title{Download docs from google docs easily}
\usage{
gdocs_download(file_id, destfile)
}
\arguments{
\item{file_id}{google file id (see the url)}

\item{destfile}{destination file}
}
\description{
Download docs from google docs easily  (given file_id and destfile)
}
