% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{gdrive_download}
\alias{gdrive_download}
\title{Download docs from Google Drive easily}
\source{
https://stackoverflow.com/questions/37453841
}
\usage{
gdrive_download(file_id, destfile)
}
\arguments{
\item{file_id}{google file id (see the url)}

\item{destfile}{destination file}
}
\description{
Download docs from Google Drive easily  (given file_id and destfile)
}
