% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{gender_import}
\alias{gender_import}
\title{Import an italian or english male/female character variable and
make a factor}
\usage{
gender_import(x, labels = c("M", "F"))
}
\arguments{
\item{x}{character the variable containing gender information (M/F)}

\item{labels}{labels to be applied to the factor}
}
\description{
Import an italian or english male/female character variable and
make a factor
}
