% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{get_preprocessors}
\alias{get_preprocessors}
\title{make a list of suitable preprocessors for omonimous argument in
preprocess_data}
\usage{
get_preprocessors(f = NULL)
}
\arguments{
\item{f}{xlsx file path (one sheet per data.frame, all are
imported); the name of the sheet should be the name of the
data.frame the preprocessors refers to; the first column (of
each sheet) is a variable name; the other colnames are function
names to be used with each variable marked. For each row
(variable) put a "x" (btw '!is.na') is used to mark the
variable with the proper coercer}
}
\description{
make a list of data preprocessors suitable for preprocessors
argument in preprocess_data given an xlsx file properly constructed
}
\examples{
\dontrun{
   proprocs <- get_preprocessors('./data/preprocessors.xlsx')
   db_clean <- preprocess_data(db, preprocs$db)
}
}
