% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_family.R
\name{group_f}
\alias{group_f}
\title{apply a function by group and return results of the same length to
the data}
\usage{
group_f(x, group, f, ...)
}
\arguments{
\item{x}{same as tapply}

\item{group}{a factor}

\item{f}{function to be applied}

\item{...}{optional arguments passed to the function}
}
\description{
apply a function by group and return results of the same length to
the data
}
\examples{
set.seed(123)
data <- data.frame(y = 1:10, g = gl(2,5, labels = c("1-5", "6-10")))
data <- data[sample(nrow(data)), ]
data$mean  <- with(data, group_f(y, g, mean))
data$min  <- with(data, group_f(y, g, min))
data$max  <- with(data, group_f(y, g, max))

}
