% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{importer}
\alias{importer}
\title{Import massively all the data from a xlsx dataset or a directory of
text file}
\usage{
importer(x, xlsx_params = list(), dir_params = list(header = TRUE,
  stringsAsFactors = FALSE, sep = ";", dec = ".", quote = "\\"", fill = TRUE))
}
\arguments{
\item{x}{char with a path to a directory or a xlsx}

\item{xlsx_params}{parameters passed to read.xlsx_alls}

\item{dir_params}{further arguments passed to read.table_dir}
}
\description{
This is a dispatcher for read.table_dir or read.xlsx_all
}
