% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%in_range\%}
\alias{\%in_range\%}
\title{Check if in range}
\usage{
x \%in_range\% y
}
\arguments{
\item{x}{a vector}

\item{y}{a vector of 2 values}
}
\description{
Check if in range
}
\examples{
x <- 1:10
y <- c(4L, 7L)
x \%in_range\% y
}
