% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inches.R
\name{inches}
\alias{inches}
\title{Translates from cm (centimeters) to inches.}
\usage{
inches(cm)
}
\arguments{
\item{cm}{Numeric vector of centimeter.}
}
\value{
A vector with inches.
}
\description{
Translates from cm (centimeters) to inches.
}
