% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythontex_utils.R
\name{include_figure}
\alias{include_figure}
\title{include a figure/includegraphics from a file in Latex code}
\usage{
include_figure(path, label, caption, scale = 1)
}
\arguments{
\item{path}{path to the pdf file}

\item{label}{label of the latex figure}

\item{caption}{caption of the latex figure}

\item{scale}{used for re-dimensioning (defaul = 1)}
}
\description{
include the figure/includegraphics from a file in Latex code
(for an image produced in R pythontex)
}
\examples{

include_figure(path_sans_ext = "/tmp/test",
             label = "test_r",
             caption = 'test R figure')

}
