% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{is.not_null_vector}
\alias{is.not_null_vector}
\title{Vector testing for humans (tm)}
\usage{
is.not_null_vector(x)
}
\arguments{
\item{x}{object to be tested}
}
\description{
returns TRUE only if x is atomic, not NULL, and with NULL
dim. It should be useful in function programming, in order to check
parameter (are they vector in the common sense meaning of the word?)
}
\examples{
is.not_null_vector(logical())
is.not_null_vector(integer())
is.not_null_vector(numeric())
is.not_null_vector(complex())
is.not_null_vector(character())
is.not_null_vector(raw())
is.not_null_vector(NULL)
is.not_null_vector(matrix(1:10, 2))
is.not_null_vector(list())        
is.not_null_vector(expression())
is.not_null_vector(pairlist())  
is.not_null_vector(pairlist(1))
}

