% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is.percentage}
\alias{is.percentage}
\title{Test if a vector represents a percentage variable}
\usage{
is.percentage(x)
}
\arguments{
\item{x}{vector to be tested}
}
\description{
Test if a vector represents a percentage variable
}
