% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is.qualitative}
\alias{is.qualitative}
\title{Test if a vector represents a qualitative variable}
\usage{
is.qualitative(x)
}
\arguments{
\item{x}{vector to be tested}
}
\description{
Test if a vector represents a qualitative variable
}

