% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is.quantitative}
\alias{is.quantitative}
\title{Test if a vector represents a quantitative variable}
\usage{
is.quantitative(x)
}
\arguments{
\item{x}{vector to be tested}
}
\description{
Test if a vector represents a quantitative variable
}

