% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unique.R
\name{is.unique}
\alias{is.unique}
\title{Utility function to check whether an object is unique}
\usage{
is.unique(x = NULL)
}
\arguments{
\item{x}{the object}
}
\description{
Utility function to check whether an object is unique
}
