% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is.vector2}
\alias{is.vector2}
\title{Vector testing for humans (tm)}
\usage{
is.vector2(x)
}
\arguments{
\item{x}{object to be tested}
}
\description{
returns TRUE only if x is atomic, not NULL, and with NULL
dim. It should be useful in function programming, in order to check
parameter (are they vector in the common sense meaning of the word?)
}
\examples{
elem_list <- list(logical = logical(),
                  integer = integer(),
                  numeric = numeric(),
                  complex = complex(),
                  char = character(),
                  raw = raw(),
                  null = NULL,
                  matrix = matrix(1:10, 2),
                  list = list(),
                  expression = expression(),
                  pairlist = pairlist(),  
                  pairlist1 = pairlist(1))
lapply(elem_list, is.vector2)
}

