% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{is.unique}
\alias{is.unique}
\title{check if a object is unique}
\usage{
is.unique(x = NULL)
}
\arguments{
\item{x}{a object that can be passed to \code{\link{duplicated}}}
}
\description{
check if a object is unique; this is a simple wrapper useful for code
readability
}

