% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_utils.R
\name{knitr_inline}
\alias{knitr_inline}
\title{knitr default pretty printer}
\usage{
knitr_inline(x)
}
\arguments{
\item{x}{R object to be printed}
}
\description{
knitr default pretty printer
}
\examples{

## library(knitr)
## opts_chunk$set("engine" = "R")
## knit_hooks$set(inline = lbmisc::knitr_inline)

}
