% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_utils.R
\name{latex_ci}
\alias{latex_ci}
\title{pretty print confidence interval from a vector}
\usage{
latex_ci(x, type = c("num", "int", "perc"), measurement_unit = "",
  digits = 1, conf_lev_string = "95\\\\\% CI",
  style = c("brack_est_ci_brack", "est_brack_ci_brack"),
  display_brackets = TRUE)
}
\arguments{
\item{x}{the length-3 vector}

\item{type}{is it a numerics, integers or percentages vector}

\item{measurement_unit}{used string for measurement unit}

\item{digits}{number of digits displayed}

\item{conf_lev_string}{used string for confidence level}

\item{style}{\code{brack_est_ci_brack} puts estimate and ci between
brackets, while \code{est_brack_ci_brack} put the estimate out of them}

\item{display_brackets}{whether to print bracket or not}
}
\description{
pretty print a confidence interval from a vector of three values
(first: estimate, second: lower ci, third: upper ci) in a latex
document
}
