% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythontex_utils.R
\name{latex_figure}
\alias{latex_figure}
\title{print the figure/includegraphics Latex code}
\usage{
latex_figure(path_sans_ext, label, caption, scale = 1)
}
\arguments{
\item{path_sans_ext}{path to the pdf file without extension (".pdf")}

\item{label}{label of the latex figure}

\item{caption}{caption of the latex figure}

\item{scale}{used for re-dimensioning (defaul = 1)}
}
\description{
print the figure includegraphics Latex code for
an image produced in R in pythontex
}
\examples{

latex_figure(file_sans_ext = "/tmp/test",
            label = "test_r",
            caption = 'test R include')

}
