% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{list2globalenv}
\alias{list2globalenv}
\title{Save a named list (eg a list of data.frames from lbmisc::import) in 
the global environment}
\usage{
list2globalenv(x, ...)
}
\arguments{
\item{x}{a named list passed to list2env}

\item{...}{other parameters passed to list2env}
}
\description{
Save a named list (eg a list of data.frames from lbmisc::import) in 
the global environment
}
