% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_letter.R
\name{margin_letter}
\alias{margin_letter}
\title{Function to draw a letter in a figure angle, after having plotted the figure}
\usage{
margin_letter(
  l = NULL,
  pos = c("topleft", "topright", "bottomleft", "bottomright", "coords"),
  cex = par("cex") * 2,
  xdr = 1/10,
  ydr = 1/10,
  coords = c(0, 0),
  ...
)
}
\arguments{
\item{l}{text}

\item{pos}{location of the letter, use 'coords' for custom non
automatic placement}

\item{cex}{cex parameter}

\item{xdr}{distance from the angle (x coordinates) as percentage of x range}

\item{ydr}{distance from the angle (y coordinates) as percentage of x range}

\item{coords}{c(x, y) vector of coordinates}

\item{...}{further parameters passed to}
}
\description{
Useful for multi-graph figures (eg with mfrow)
}
\examples{
plot(rnorm(100))
margin_letter("A", "topleft")
margin_letter("B", "topright", xdr = 1/50) ## decrease distance from the angle, otherwise it's printed out of screen
margin_letter("C", "bottomleft")
margin_letter("D", "bottomright", xdr = 1/50)
}
