% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu2.R
\name{menu2}
\alias{menu2}
\title{menu-like function for interactive or non-interactive sections
which allows multiple choices as well}
\usage{
menu2(
  choices,
  title = NULL,
  multiple = FALSE,
  return = c("values", "indexes"),
  strict = FALSE
)
}
\arguments{
\item{title}{optional}

\item{multiple}{can more than one item be selected?}

\item{return}{what to return values (selected choiches given, by
default), or indexes. If only 0 is selected (to exit), NA is
returned}

\item{strict}{allow only selectable index to be choosen}

\item{choiches}{vector of possible choiches}
}
\description{
return NA if the user inserted the usual 0 to exit
}
