% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu_ni.R
\name{menu_ni}
\alias{menu_ni}
\title{menu-like functions for non interactive sections}
\usage{
menu_ni(choices, title = NULL, allow_multiple = FALSE)
}
\arguments{
\item{title}{give a title?}

\item{allow_multiple}{allows multiple choices}

\item{choiches}{vector of possible choiches}
}
\description{
menu-like functions for non interactive sections
}
