% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge2.R
\name{merge2.data.frame}
\alias{merge2.data.frame}
\title{A verbose merge}
\usage{
\method{merge2}{data.frame}(
  x,
  y,
  id_dupl = c("x", "y"),
  id_not_found = c("y_nin_x"),
  by = intersect(names(x), names(y)),
  by.x = by,
  by.y = by,
  all = FALSE,
  all.x = all,
  all.y = all,
  sort = TRUE,
  suffixes = c(".x", ".y"),
  no.dups = TRUE,
  incomparables = NULL,
  ...
)
}
\arguments{
\item{x}{same as in \code{merge}}

\item{y}{same as in merge}

\item{id_dupl}{character, warning if duplicated id are found for
the selected data.frame}

\item{id_not_found}{message if some A id are not found in B (eg
x_nin_y or y_nin_x)}

\item{by}{same as in merge}

\item{by.x}{same as in merge}

\item{by.y}{same as in merge}

\item{all}{same as in merge}

\item{all.x}{same as in merge}

\item{all.y}{same as in merge}

\item{sort}{same as in merge}

\item{suffixes}{same as in merge}

\item{no.dups}{same as in merge}

\item{incomparables}{same as in merge}

\item{...}{same as in merge}
}
\description{
A verbose merge
}
