% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge2.R
\name{merge2}
\alias{merge2}
\title{A verbose merge}
\usage{
merge2(x, y, ...)
}
\arguments{
\item{x}{first element}

\item{y}{second element}

\item{...}{further arguments passed to methods}
}
\description{
A verbose merge
}
