% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_utils.R
\docType{data}
\name{mesi.abb}
\alias{mesi.abb}
\title{Abbreviated months in italian, the easy way}
\format{An object of class \code{character} of length 12.}
\usage{
mesi.abb
}
\description{
Abbreviated months in italian, the easy way
}
\keyword{datasets}
