% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_utils.R
\name{month}
\alias{month}
\title{Extract numeric months of the year from a Date, POSIXct or POSIXlt vector.}
\usage{
month(x, string = TRUE, abbreviate = FALSE)
}
\arguments{
\item{x}{Date, POSIXct or POSIXlt vector}

\item{string}{Logical, if TRUE months string is returned, otherwise numeric}

\item{abbreviate}{logical. Should the months in string format be
abbreviated?}
}
\value{
A vector with extracted months.
}
\description{
Extract numeric months of the year from a Date, POSIXct or POSIXlt vector.
}
\examples{

Sys.Date()
month(Sys.Date())
month(Sys.Date(), string = FALSE)
month(Sys.Date(), string = TRUE, abbreviate = TRUE)
## from base package
month(Sys.Date())
month(as.POSIXct(Sys.Date()))
month(as.POSIXlt(Sys.Date()))


}
