% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_utils.R
\name{NA_plot}
\alias{NA_plot}
\title{plot patterns of missing values}
\usage{
NA_plot(x, ...)
}
\arguments{
\item{x}{a data.frame or a matrix}

\item{...}{other options passed to proper methods}
}
\description{
plot patterns of missing values
}
\examples{

## NA_plot(airquality)
## NA_plot(. ~ Month, data = airquality)
}
