% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_utils.R
\name{NA_remove}
\alias{NA_remove}
\title{remove NA and make a message with number of units deleted}
\usage{
NA_remove(x, verbose = TRUE)
}
\arguments{
\item{x}{something handled by \code{stats::na.omit}}

\item{verbose}{logical: if \code{TRUE} raise a message if any row is
deleted}
}
\description{
remove NA and make a message with number of units deleted
}

