% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_utils.R
\name{NA_report}
\alias{NA_report}
\title{Report NA variables per case}
\usage{
NA_report(x, id_var = NULL)
}
\arguments{
\item{x}{data.frame to be analyzed}

\item{id_var}{variable to be used as id}
}
\description{
Report NA variables per case
}
\examples{
test <- data.frame(
 name =  c("john","mary","gregor"),
 x = c(1,NA,NA),
 y = c(NA,1,1),
 z = c(1,1,NA))
test2 <- test
test2$surname <- c("doe", "foo", "bar")
test2 <- test2[c("name","surname","x","y","z")]

test
NA_report(test, "name")

test2
NA_report(test2, c("name", "surname"))
}
