% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{names_list}
\alias{names_list}
\title{Print the names of x as list for easy paste/copy in excel}
\usage{
names_list(x)
}
\arguments{
\item{x}{the object with names}
}
\description{
Print the names of x as list for easy paste/copy in excel
}
