% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Inverse Value Matching}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\description{
Complement of \code{\%in\%}. Returns the elements of \code{x}
that are not in \code{y}.
}

