% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{nosi_import}
\alias{nosi_import}
\title{Import an italian or english no/si-yes character variable and make
a factor}
\usage{
nosi_import(x, labels = c("No", "Yes"))
}
\arguments{
\item{x}{character the variable}

\item{labels}{labels to be applied to the factor}
}
\description{
Import an italian or english no/si-yes character variable and make
a factor
}
