% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pddf.R
\name{pddf}
\alias{pddf}
\title{Generate an ascii representation of a dataframe suitable for pandas}
\usage{
pddf(df, outfile = "", dfname = "df")
}
\arguments{
\item{df}{a data.frame}

\item{outfile}{a .py file path or the dataframe will be printed on
stdout}

\item{dfname}{name given to the pd.DataFrame}
}
\description{
It's a dput for python pandas dataframe
}
\examples{
pddf(Indometh)
pddf(data.frame("a" = pi))

}
