% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pii_erase.R
\name{pii_cleaner}
\alias{pii_cleaner}
\title{Import multiple dataset, remove PII and save anonymized}
\usage{
pii_cleaner(input = tcltk::tk_choose.files(caption =
  "Select DATA FILES to be imported and anonymized"), outfile = NULL, ...)
}
\arguments{
\item{input}{importer x param}

\item{outfile}{exporter x param}
}
\description{
Import multiple dataset, remove PII and save anonymized
}
