% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pii_erase.R
\name{pii_erase}
\alias{pii_erase}
\title{Remove PII (Personally identifiable information) from a data.frame}
\usage{
pii_erase(x, quiet = FALSE)
}
\arguments{
\item{x}{data.frame to be anonymized}

\item{quiet}{if FALSE (by default) emit a message with erased variables}
}
\description{
Remove PII (Personally identifiable information) from a data.frame
}
