% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{plot_3d}
\alias{plot_3d}
\title{Plot a z = f(x,y) function}
\usage{
plot_3d(f = NULL, xrange = c(-1, 1), yrange = c(-1, 1),
  points = 50, contour = FALSE, ...)
}
\arguments{
\item{f}{a f(x,y) function}

\item{xrange}{x domain}

\item{yrange}{y domain}

\item{points}{number of points}

\item{contour}{if TRUE make a contour plot instead of 3d plot}
}
\description{
Plot a z = f(x,y) function
}
\examples{
fun <- function(x, y) x^2 + y^2
par(mfrow = c(1,2))
plot_3d(fun)
plot_3d(fun, contour = TRUE)

}
