% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{plot_fun}
\alias{plot_fun}
\title{Plot a mathematical function}
\usage{
plot_fun(domain = c(0, 5), fun = function(x) x + 1,
  cartesian_plane = TRUE)
}
\arguments{
\item{domain}{a vector of 2 elements}

\item{fun}{a function}

\item{cartesian_plane}{wheter to add a cartesian plane}
}
\value{
Nothing. As a side effect the plot of \code{pch}.
}
\description{
Plot a 2d mathematical function y = f(x)
}
\examples{

plot_fun()

}
