% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{plot_pfun}
\alias{plot_pfun}
\title{Plot a parametric mathematical function}
\usage{
plot_pfun(f = NULL, from = NULL, to = NULL, points = 100,
  col = "black", cartesian_plane = TRUE,
  cartesian_plane_col = "black", grid_at_x = NULL, grid_at_y = NULL,
  add = FALSE, ...)
}
\arguments{
\item{f}{a named list of function}

\item{from}{domain (plotting) starting point}

\item{to}{domain (plotting) ending point}

\item{points}{number of points}

\item{col}{the function color}

\item{cartesian_plane}{wheter to add a cartesian plane ...}

\item{cartesian_plane_col}{... and its color}

\item{grid_at_x}{at_x param for add_grid}

\item{grid_at_y}{at_y param for add_grid}

\item{add}{if TRUE the function is added within an existing graph}

\item{...}{other params given to plot}
}
\value{
Nothing. As a side effect the plot of \code{pch}.
}
\description{
Plot a 2d or 3d parametric mathematical function f(t) = (x(t), y(t), z(t))
}
\examples{
pfun <- list(x = function(t) 2 * cos(t), y = function(t) 3 * sin(t))
plot_pfun(pfun, from = 0, to = pi)
plot_pfun(pfun, from = 0, to = 2*pi)

pfun2 <- list(x = function(t) 1 * cos(t),
              y = function(t) 1 * sin(t),
              z = function(t) 1 * t)
plot_pfun(pfun2, from = 0, to = 100, points = 1000)

}
