% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{preprocess data}
\usage{
preprocess_data(x, types, mr_name = "mr", mr_fun = smarty_mr_splitter, ...)
}
\arguments{
\item{x}{the data.frame}

\item{types}{a named list: names = function names to be applied, 
content is variable to be treated with the function}

\item{mr_name}{name of types associated used for multiple response}

\item{mr_fun}{function used for multiple response which creates a data.frame
of dummies given a single vector of multiple responses (eg 
separated by "|")}

\item{...}{argument passed to mr_fun}
}
\description{
preprocess data according to variable name - function couple directives
}
\examples{
data_char <- function(d) as.Date(d, format = '\%d/\%m/\%Y')
noyes <- function(x) factor(x, levels = c('NO', 'SI'),
                            labels = c('No', 'Yes'))
test_df <- data.frame('a_date'    = c('2017-01-01', '2015-01-01'),
                      'gender'    = c('m', 'f'),
                      'female'    = c('NO', 'SI'),
                      'educ_lev'  = c('diploma', 'degree'),
                      'birth'     = c("11/08/1927", "24/05/1935"),
                      'interests' = c('reading|travel|science',
                                      'reading|science|cinema|tv'),
                      stringsAsFactors = FALSE)
types <- list("as.Date" = 'a_date',
              "factor" = c("gender", "educ_lev"),
              "data_char" = 'birth',
              "noyes" = 'female',
              "mr" = "interests")
str(test_df)
preproc <- preprocess_data(test_df, types)
str(preproc)
}
