% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_utils.R
\name{preprocess_varnames}
\alias{preprocess_varnames}
\title{Preprocess data.frame variable names}
\usage{
preprocess_varnames(x = NULL, trim = NULL, dump_rev = TRUE)
}
\arguments{
\item{x}{a data.frame or its (col)names}

\item{trim}{character length of trimming. If \code{NULL} (default)
trimming is disabled.}

\item{dump_rev}{print a matrix with the reverse renaming. It can be used
with \code{recode} (back to original names?) or
\code{comment_df} (if original names are to be used as
comments)}
}
\description{
Function to preprocess variable names useful for a data.frame.  If
a char is given a char is correct char is given back, if a
data.frame the same data.frame with corrected names.
}
\details{
This function was created to make automatic variable name creation
from Excel files obtained by others.
}
