% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_pval.R
\name{pretty_pval}
\alias{pretty_pval}
\title{Pretty print for p-values}
\usage{
pretty_pval(pvalue, digits = 3L, space = FALSE, equal = FALSE)
}
\arguments{
\item{pvalue}{A numeric vector of p-values.}

\item{digits}{number of digits returned (3 by default).}

\item{space}{Logical specifying whether a space should be inserted between
number and operator (= or <, default to FALSE).}

\item{equal}{add = where needed?}
}
\value{
The function return a string with the pretty printed p-values.
}
\description{
Pretty print for p-values.
}
\examples{

pval1 <- c(3, NA, 1e-01, 1e-02, 1e-03, 1e-04, 1e-05)
pretty_pval(pval1, space = TRUE)

}

