% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_utils.R
\name{pretty_text}
\alias{pretty_text}
\title{Prepare a dataframe variable names for printing}
\usage{
pretty_text(x, what = c("rm_", "upcase_first"))
}
\arguments{
\item{x}{a data.frame}
}
\description{
eg remove _
}
