% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{print_unique_values}
\alias{print_unique_values}
\title{Print the unique values of the elements composing x}
\usage{
print_unique_values(x)
}
\arguments{
\item{x}{the object (tipically data.frame or list)}
}
\description{
Print the unique values of the elements composing x
}
