% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_utils.R
\name{pval_stars}
\alias{pval_stars}
\title{Pretty print for p-values}
\usage{
pval_stars(p = NULL, less_05 = TRUE)
}
\arguments{
\item{p}{A numeric vector of p-values.}

\item{less_05}{if TRUE (default) return stars only for p < 0.05}
}
\value{
The function returns a character vector with stars related
        to p-value.
}
\description{
Pretty print for p-values.
}
\examples{

pval1 <- c(NA, 0, 0.005, 0.01, 0.025, 0.05, 0.07, 0.5, 1, 3)
cbind(pval1, pval_stars(pval1))

}
