% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.table_dir}
\alias{read.table_dir}
\title{Apply read.table to all the files with a given extension in a given
directory}
\usage{
read.table_dir(dir = "data/dataset/", ext = "csv", ret = c("list",
  "assign"), ...)
}
\arguments{
\item{dir}{a directory file to be readed with read.table}

\item{ext}{which file extension files have to have to be readed}

\item{ret}{either assign or a list}

\item{...}{other options passed to read.table}
}
\description{
Apply read.table to all the files with a given extension in a given
directory
}
