% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.tables}
\alias{read.tables}
\title{Apply read.table to all the files given as parameter and return a
list}
\usage{
read.tables(f, name_gen = default_name_gen, verbose = TRUE, ...)
}
\arguments{
\item{name_gen}{function that return data.frame name given the path
to his file}

\item{verbose}{if TRUE (by default) print info about progression
(file imported)}

\item{...}{options passed to read.table}
}
\description{
Apply read.table to all the files given as parameter and return a
list
}
