% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.xlsx_alls}
\alias{read.xlsx_alls}
\title{Read all sheets from an xlsx}
\usage{
read.xlsx_alls(f = NULL, ret = c("list", "assign"), ...)
}
\arguments{
\item{f}{xlsx file}

\item{ret}{either list or assign}

\item{...}{other options passed to openxlsx::read.xlsx}
}
\description{
Read all sheets from an xlsx; return them as a list or assign in the
calling environment
}
