% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{recist_import}
\alias{recist_import}
\title{ease recist codes import}
\usage{
recist_import(x)
}
\arguments{
\item{x}{character recist code such as CR, CP etc}
}
\description{
ease recist codes import
}
