% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{recode}
\alias{recode}
\title{Vector recode utility}
\usage{
recode(x = NULL, rec = NULL)
}
\arguments{
\item{x}{Vector to be recoded.}

\item{rec}{matrix containing 2 columns; first is "from", second is
"to".}
}
\value{
A vector with new codes (if/where modified).
}
\description{
Yet another recode utility
}

