% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_NA.R
\name{remove_NA}
\alias{remove_NA}
\title{remove NA and make a message with number of units deleted}
\usage{
remove_NA(x, verbose = TRUE)
}
\arguments{
\item{x}{something handled by \code{stats::na.omit}}

\item{verbose}{logical: if \code{TRUE} raise a message if any row is
deleted}
}
\description{
remove NA and make a message with number of units deleted
}

