% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_splitter.R
\name{replace_mr_with_dummies}
\alias{replace_mr_with_dummies}
\title{Replace smarty multiple responses with proper dummies}
\usage{
replace_mr_with_dummies(x = NULL, variables = NULL, drop = FALSE)
}
\arguments{
\item{x}{data.frame of interest}

\item{variables}{mr variables to be replaced}

\item{drop}{remove original mr variable once dummies have been added?}
}
\description{
Replace smarty multiple responses with proper dummies
}
