% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_utils.R
\name{rm_accents}
\alias{rm_accents}
\title{Remove accents from a character vector}
\usage{
rm_accents(string)
}
\arguments{
\item{string}{a character vector}
}
\description{
Remove accents from a character.
}

