% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_utils.R
\name{rm_spaces}
\alias{rm_spaces}
\title{Remove spaces from a character vector}
\usage{
rm_spaces(string, all = FALSE)
}
\arguments{
\item{string}{a character vector}

\item{all}{logical, if TRUE remove all blank spaces, otherwise rm only
duplicated and leading/trailing ones.}
}
\description{
This function removes spaces (all leading and trailing ones,
and unique in-between) from a character vector.
}
\examples{
test <- c("  test ", "  mediterranean  sea  ")
rm_spaces(test)
}
