% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_utils.R
\name{rm_unprintable_chars}
\alias{rm_unprintable_chars}
\title{Remove unprintable chars}
\usage{
rm_unprintable_chars(string)
}
\arguments{
\item{string}{a character vector}
}
\description{
Remove unprintable chars
}

