% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{rownames2col1}
\alias{rownames2col1}
\title{add rownames as first colum of a dataframe}
\usage{
rownames2col1(x, nm = "x")
}
\arguments{
\item{x}{data.frame}

\item{nm}{name to give to the first columnù}
}
\description{
add rownames as first colum of a dataframe: useful for xtp_atwb where rownames
are to be saved
}
\examples{
\dontrun{
xtp_atwb(rownames2col1(Indometh),
         wb = wb, sheet = 'Indometh', label = 'tab:indometh',
         caption = 'Indometh dataset')
}

}
