% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments_text.R
\name{segments_text}
\alias{segments_text}
\title{Display a line with text above in a figure}
\usage{
segments_text(text, pos = c("above", "below"), x0 = 1, x1 = 2, y,
  lty = "dotted")
}
\arguments{
\item{text}{text displayed}

\item{pos}{text position regarding the line}

\item{x0}{line coordinates x_0 (by default = 1, to ease boxplots)}

\item{x1}{line coordinates x_1 (by default = 2, to ease boxplots)}

\item{y}{line coordinates y (horizontal line only)}

\item{lty}{line lty}
}
\description{
especially useful for boxplot (p-value comparison of groups)
}
