% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setRownames.R
\name{setRownames}
\alias{setRownames}
\title{set rownames of an object and return it}
\usage{
setRownames(object = nm, nm)
}
\arguments{
\item{object}{an object with rows (where rownames works)}

\item{nm}{rownames to be assigned}
}
\description{
set rownames of an object and return it
}
