% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{show_col}
\alias{show_col}
\title{Show R colors for graphics and grid package}
\usage{
show_col(package = c("graphics", "grid"), bg = "white", cex = 0.75,
  srt.rot = 30)
}
\arguments{
\item{package}{charachter. Which package's color to show}

\item{bg}{charachter. Background color}

\item{cex}{numeric \code{cex}}

\item{srt.rot}{degree inclination}
}
\value{
Nothing. As a side effect the plot of colors.
}
\description{
Show R colors for graphics and grid package (simple wrapper around functions
from demo(colors).
}
\examples{

show_col()

}

