% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_if.R
\name{show_if}
\alias{show_if}
\title{show a chunk of a report if a logical condition is met}
\usage{
show_if(x)
}
\arguments{
\item{x}{a logical value}
}
\description{
start showing a chunk of a report if a logical condition is met; to
be used in a results = 'asis' chunk of Sweave/knitr
}
\examples{

eval_old_stuff <- FALSE
show_if(eval_old_stuff)

}
