% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{show_pch}
\alias{show_pch}
\title{Show \code{pch}}
\usage{
show_pch(
  extras = c("*", ".", "0", "+", "#"),
  cex = 3,
  col = "red3",
  bg = "gold",
  cextext = 1.2
)
}
\arguments{
\item{extras}{charachter. Further symbols to be plotted}

\item{cex}{numeric \code{cex}}

\item{col}{color}

\item{bg}{character. Background color}

\item{cextext}{\code{cex} for text}
}
\value{
Nothing. As a side effect the plot of \code{pch}.
}
\description{
Show \code{pch} values.
}
\examples{

show_pch()

}
