% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smarty_mr_splitter.R
\name{smarty_mr_splitter}
\alias{smarty_mr_splitter}
\title{Smarty Multiple Response Splitter}
\usage{
smarty_mr_splitter(x, spl_char = "|", categs = NULL,
  add_count = FALSE)
}
\arguments{
\item{x}{variable to be splitted}

\item{spl_char}{splitting character}

\item{categs}{character vector listing the categories to be
considered and/or they order; if \code{NULL} (default) all
categories given in x will be used (in alphabetic order)}

\item{add_count}{add a variable for
considered and/or they order; if \code{NULL} (default) all
categories given in x will be used (in alphabetic order)}
}
\description{
This function split a variable containing Smarty's output for
a multiple response question in a dummy data.frame, to be
used with \code{\link{cbind}}.
}
