% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss_utils.R
\name{spss_datetime}
\alias{spss_datetime}
\title{Convert datetime values from SPSS}
\usage{
spss_datetime(x, coercer = as.Date)
}
\arguments{
\item{x}{number of seconds since 1582-10-14}

\item{coercer}{function to use as a coercer given the desidered output}
}
\description{
Convert the number of seconds since 1582-10-14 available in SPSS datetime
to a format specified by coercer
}
