% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2df.R
\name{table2df}
\alias{table2df}
\title{make a data.frame from a table of counts}
\usage{
table2df(x)
}
\arguments{
\item{x}{a frequency table (without margins)}
}
\description{
make a data.frame from a table of counts
}
