% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_utils.R
\name{to_00_char}
\alias{to_00_char}
\title{Transform an integer vector to a character prefixed by an arbitrary
number of 0 digits}
\usage{
to_00_char(x, clen = NULL)
}
\arguments{
\item{x}{a numeric vector}

\item{clen}{length of final character vector}
}
\description{
Transform an integer vector to a character prefixed by an arbitrary
number of 0 digits
}
\examples{
to_00_char(c(1L, 3L, 20L), 3)
to_00_char(c(1L, 3L, 20L), 2)
\dontrun{
to_00_char(c(1L, 3L, 20L), 2)
}
}

