% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unoconv.R
\name{unoconv}
\alias{unoconv}
\title{Unoconv file converter from R}
\usage{
unoconv(..., format = stop("Format needed"))
}
\arguments{
\item{...}{file path to be exported}

\item{format}{file format}
}
\description{
Unoconv file converter from R
}
