% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upcase.R
\name{upcase}
\alias{upcase}
\title{Capitalize words in a vector of characters}
\usage{
upcase(s, method = c("sentence", "word", "strict"))
}
\arguments{
\item{s}{vector of characters}

\item{method}{either sentence (first letter in each element of the
vector), word (first letter in each word of the sentence) or strict (
first letter in each word of the sentence)}
}
\description{
Capitalize words in a vector of characters
}
\examples{
upcase(c("using AIC", "for model selection"))
upcase(c("using AIC", "for model selection"), method = 'word')
upcase(c("using AIC", "for model selection"), method = 'strict')
}
