% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{update_packages}
\alias{update_packages}
\title{update package shortcut}
\usage{
update_packages(ask = FALSE)
}
\arguments{
\item{ask}{same as \code{\link{update.packages}}}
}
\description{
update package shortcut with my defaults
}
