% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose_coerce.R
\name{verbose_coerce}
\alias{verbose_coerce}
\title{Coerce/convert data verbosely if NA is introduced by coercion}
\usage{
verbose_coerce(x, coercer, xname = NULL, varname_msg = TRUE)
}
\arguments{
\item{x}{vector to be converted}

\item{coercer}{a function (tipically as.*)}

\item{xname}{string used for progress (varname) printing}

\item{varname_msg}{logical: print progress (varname)?}
}
\description{
Coerce/convert a given x using the coercer function; if a
warning is raised and a difference in NA cardinality is 
present a report of NA produced by coercion is printed
}
\examples{
\dontrun{verbose_coerce(c("a", 2L, 3L), as.integer)}

}
