% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_family.R
\name{verbose_lapply}
\alias{verbose_lapply}
\title{Print names of the list elements as progress while simulating
lapply to a list}
\usage{
verbose_lapply(x, FUN, ...)
}
\arguments{
\item{x}{list}

\item{FUN}{function to be applied}

\item{...}{further parameters passed to FUN}
}
\description{
Print names of the list elements as progress while simulating
lapply to a list
}
