% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{Alternative to View that uses openXL}
\usage{
view(..., freezePane = TRUE, autoColWidth = TRUE)
}
\arguments{
\item{...}{writeData handled objects}

\item{freezePane}{apply freezePane at 'A2', for each sheet?}

\item{autoColWidth}{apply auto column widths to all sheets?}
}
\description{
Alternative to \code{\link{View}} that invokes a spreasheet viewer on
openxlsx's \code{\link{writeData}} handled objects.
}
\examples{
\dontrun{
view(Indometh, iris)
}
}
