% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{wb_to_xl}
\alias{wb_to_xl}
\title{Save an openxlsx workbook in an excel format quickly}
\usage{
wb_to_xl(wb, file = NULL)
}
\arguments{
\item{wb}{openxlsx workbook}

\item{file}{xls destination file}
}
\description{
Save an openxlsx workbook in an excel format quickly
}
\examples{
\dontrun{
wb = openxlsx::createWorkbook()
add_to_wb(wb = wb, sheet = 'Indometh', x = Indometh)
wb_to_xl(wb = wb, file = '/tmp/wb_to_xl.xls')
}
}
