% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{wb_to_xls}
\alias{wb_to_xls}
\title{A function to save an openxlsx workbook in xls format quickly}
\usage{
wb_to_xls(wb, destfile = NULL)
}
\arguments{
\item{wb}{openxlsx workbook}

\item{destfile}{xls destination file}
}
\description{
A function to export an openxlsx workbook oin xls format quickly
}

