% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_lines}
\alias{write_lines}
\title{Write lines to a text file, the quick way}
\usage{
write_lines(x, f)
}
\arguments{
\item{x}{lines (characters)}

\item{f}{path to the file}
}
\description{
Write lines to a text file, the quick way
}
