% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtile.R
\name{xtile}
\alias{xtile}
\title{Cut numeric x in quantile}
\usage{
xtile(x, n)
}
\arguments{
\item{x}{numeric vector}

\item{n}{number of quantiles}
}
\value{
A factor vector.
}
\description{
Cut numeric x in quantile
}
\examples{

tertiles <- xtile(rnorm(1000),3)
table(tertiles)

}
