% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{xtp_atwb}
\alias{xtp_atwb}
\title{print as xtable and add to workbook}
\usage{
xtp_atwb(
  x,
  wb = NULL,
  sheet = "",
  label = "",
  caption = "",
  xtable_par = list(digits = 3),
  print_xtable_par = list(include.rownames = FALSE),
  add_to_wb_par = list(rowNames = FALSE)
)
}
\arguments{
\item{x}{the object}

\item{wb}{add_to_wb workbook name}

\item{sheet}{add_to_wb sheet name}

\item{label}{xtable::xtable label}

\item{caption}{xtable::xtable caption}

\item{xtable_par}{xtable::xtable further parameters}

\item{print_xtable_par}{xtable::print.xtable further parameters}

\item{add_to_wb_par}{add_to_wb further parameters}
}
\description{
print as xtable and add to workbook
}
\examples{
\dontrun{
xtp_atwb(Indometh,
         wb = wb, sheet = 'Indometh', label = 'tab:indometh',
         caption = 'Indometh dataset')
}
}
