% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_utils.R
\name{year}
\alias{year}
\title{Extract numeric years from a Date, POSIXct or POSIXlt vector.}
\usage{
year(x)
}
\arguments{
\item{x}{Date, POSIXct or POSIXlt vector}
}
\value{
A numeric vector.
}
\description{
Extract numeric years from a Date, POSIXct or POSIXlt vector.
}
\examples{

Sys.Date()
year(Sys.Date())
year(as.POSIXct(Sys.Date()))
year(as.POSIXlt(Sys.Date()))


}

