% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{add_cartesian_plane}
\alias{add_cartesian_plane}
\title{Add cartesian system to a plot}
\usage{
add_cartesian_plane(at_y = 0, at_x = 0, col = "black", lty = "solid",
  lwd = 1, ...)
}
\arguments{
\item{at_y}{y centre of the system}

\item{at_x}{x centre of the system}

\item{col}{grid color}

\item{lty}{line type}

\item{lwd}{line width}

\item{...}{further parameter passed to abline}
}
\value{
The function adds a cartesian plane to a plot using
\code{\link{abline}}.
}
\description{
Add a cartesian system to a plot using \code{\link{abline}}.
}

