% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{add_grid}
\alias{add_grid}
\title{Add grid to a plot}
\usage{
add_grid(at.x, at.y = at.x, col = "lightgray", lty = "dotted",
  lwd = par("lwd"))
}
\arguments{
\item{at.x}{x axis grid values}

\item{at.y}{y axis grid values}

\item{col}{grid color}

\item{lty}{line type}

\item{lwd}{line width}
}
\value{
The function adds a grid to a plot using \code{\link{abline}}.
}
\description{
Add grid to a plot using \code{\link{abline}}.
}

