% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argparse.R
\name{argparse}
\alias{argparse}
\title{Simple command line argument parser}
\usage{
argparse()
}
\description{
Simple command line argument parser: returns a named list with
arguments and values
}
\examples{
## Beginning of an Rscript
#!/usr/bin/Rscript --vanilla --quiet
args <- lbmisc::argparse()
}
