% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{compare_columns}
\alias{compare_columns}
\title{Compare columns progressively}
\usage{
compare_columns(db, operator = "<", row_id = NULL, short = TRUE)
}
\arguments{
\item{db}{a data.frame with ordered columns}

\item{operator}{comparison operator}

\item{row_id}{an optional row id}

\item{short}{logical, whether to include only a report (TRUE) or a matrix
with all comparisons results (FALSE)}
}
\value{
a list with raw results and a report
}
\description{
Compare columns progressively in a dataset using a specified operator,
that tells how columns should be ordered (eg by default columns should
be increasing)
}
\examples{

compare_columns(data.frame(a = c(1,2,3), b = c(0,1,4)),
                row_id = letters[1:3])

}
