% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{compare_columns}
\alias{compare_columns}
\title{Compare columns progressively}
\usage{
compare_columns(
  db = NULL,
  operator = "<",
  row_id = seq_len(nrow(db)),
  extended = TRUE
)
}
\arguments{
\item{db}{a data.frame with ordered columns}

\item{operator}{comparison operator}

\item{row_id}{an optional row id}

\item{extended}{logical if TRUE}
}
\value{
a list of checks with values or (if extended = FALSE) a
    compact data.frame with non-passed checks and variable mentioning
}
\description{
Compare columns progressively in a dataset using a specified
operator, that tells how columns should be ordered (eg by default
columns should be increasing)
}
\examples{

(data <- data.frame(id = letters[1:4],  x = c(1,2,3,4), y = c(0,1,3,NA), z = rep(0, 4)))
compare_columns(data[, -1], operator= '<')
compare_columns(data[, -1], operator= '>')
compare_columns(data[, -1], operator= '<', row_id = data$id)
compare_columns(data[, -1], operator= '<', row_id = data$id, extended = FALSE)

}
