% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_utils.R
\name{dmY_import}
\alias{dmY_import}
\title{Quick alias for importing italian csv common date formats  (dd/mm/yyyy)}
\usage{
dmY_import(x, fmt = "\%d/\%m/\%Y", ...)
}
\arguments{
\item{x}{character with date}

\item{fmt}{format given to as.Date}

\item{...}{other arguments passed to as.Date}
}
\description{
Quick alias for importing italian csv common date formats  (dd/mm/yyyy)
}
