% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{importer}
\alias{importer}
\title{Import massively all the data from a xlsx dataset or a directory of
text file}
\usage{
importer(
  p,
  xlsx_params = list(),
  text_params = list(header = TRUE, stringsAsFactors = FALSE, sep = ";", dec = ".",
    quote = "\\"", fill = TRUE)
)
}
\arguments{
\item{p}{char with a path to multiple text data files, a directory
with text data file or a single .xlsx file}

\item{xlsx_params}{parameters passed to read.xlsx_alls (they will
be commonly applied to all xlsx)}

\item{text_params}{arguments passed to read.table (they will be
commonly applied to all text files)}
}
\description{
This is a dispatcher for read.tables, read.table_dir or
read.xlsx_alls
}
