% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu2.R
\name{menu2}
\alias{menu2}
\title{menu-like function for interactive or non-interactive sections
which allows multiple choices as well}
\usage{
menu2(
  choices,
  title = NULL,
  multiple = FALSE,
  return = c("value", "index"),
  strict = FALSE
)
}
\arguments{
\item{title}{optional}

\item{multiple}{can more than one item be selected?}

\item{return}{what to return values (selected choiches given, by
default), or indexes}

\item{strict}{allow only selectable index to be choosen}

\item{choiches}{vector of possible choiches}
}
\description{
menu-like function for interactive or non-interactive sections
which allows multiple choices as well
}
