% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_utils.R
\name{NA_remove}
\alias{NA_remove}
\title{remove NA and make a message with number of units deleted}
\usage{
NA_remove(x, ...)
}
\arguments{
\item{x}{something handled by \code{stats::na.omit}}

\item{...}{other options passed to proper methods}
}
\description{
like na.omit but print a message
}
\examples{
nona <- NA_remove(data.frame(a = 1:6,
                             b = c(NA, 3, NA, 4, 5, 6),
                             c = c(rep(NA, 4), 'a', 'b')))

}
