% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pddf.R
\name{pddf}
\alias{pddf}
\title{Generate an ascii representation of a dataframe suitable for pandas}
\usage{
pddf(df)
}
\arguments{
\item{df}{a data.frame}
}
\description{
It's a dput for python pandas dataframe
}
\examples{
pddf(Indometh)
pddf(data.frame("a" = pi))

}
