% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{plot_fun}
\alias{plot_fun}
\title{Plot a mathematical function (enhanced version of curve)}
\usage{
plot_fun(f = function(x) x + 1, from = 0, to = 5, f_col = "blue",
  cartesian_plane = TRUE, cartesian_plane_col = "black", grid_at_x = NULL,
  grid_at_y = NULL, ...)
}
\arguments{
\item{f}{a function}

\item{from}{domain (plotting) starting point}

\item{to}{domain (plotting) ending point}

\item{cartesian_plane}{wheter to add a cartesian plane ...}

\item{cartesian_plane_col}{... and its color}

\item{grid_at_x}{at_x param for add_grid}

\item{grid_at_y}{at_y param for add_grid}

\item{...}{other params given to plot}

\item{f}{the function color}
}
\value{
Nothing. As a side effect the plot of \code{pch}.
}
\description{
Plot a 2d mathematical function y = f(x)
}
\examples{

plot_fun()

}
