% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_utils.R
\name{preprocess_varnames}
\alias{preprocess_varnames}
\title{Preprocess data.frame variable names}
\usage{
preprocess_varnames(varnames = NULL, trim = NULL)
}
\arguments{
\item{varnames}{names of a data.frame (or the data.frame itself)}

\item{trim}{character length of trimming. If \code{NULL}
(default) trimming is disabled.}
}
\description{
Function to preprocess variable names useful for a data.frame.
This function was created to make automatic variable name
creation from Excel files obtained by other.
}

