% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_utils.R
\name{pval_stars}
\alias{pval_stars}
\title{Pretty print for p-values}
\usage{
pval_stars(p = NULL)
}
\arguments{
\item{p}{A numeric vector of p-values.}
}
\value{
The function returns a character vector with stars related
        to p-value.
}
\description{
Pretty print for p-values.
}
\examples{

pval1 <- c(3, NA, 0.005, 0.025, 0.05, 0.07, 0.5)
pval_stars(pval1)

}
