% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.table_dir}
\alias{read.table_dir}
\title{Apply read.table to all the files with a given extension in a given
directory}
\usage{
read.table_dir(d, ...)
}
\arguments{
\item{d}{a directory file to be readed with read.table}

\item{...}{other options passed to read.tables}
}
\description{
Apply read.table to all the files with a given extension in a given
directory
}
