% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose_coerce.R
\name{verbose_coerce}
\alias{verbose_coerce}
\title{Coerce/convert data verbosely if NA is introduced by coercion}
\usage{
verbose_coerce(
  x,
  coercer,
  xname = NULL,
  varname_msg = TRUE,
  row_id = NULL,
  df_name = ""
)
}
\arguments{
\item{x}{vector to be converted}

\item{coercer}{a function (tipically as.*)}

\item{xname}{string used for progress (varname) printing}

\item{varname_msg}{logical: print progress (varname and optionally
source)?}

\item{row_id}{vector or data.frame used to identify the coerced
rows (for reporting purposes in case of issues)}

\item{df_name}{character with data.frame name for reporting
purposes (where does the variable come from?)}
}
\description{
Coerce/convert a given x using the coercer function; if a warning
is raised and a difference in NA cardinality is present a report of
NA produced by coercion is printed, with optional rows id printed
}
\examples{
\dontrun{
verbose_coerce(c("a", 2L, 3L), as.integer)
verbose_coerce(c("a", 2L, 3L), as.integer, row_id = letters[1:3])
}

}
