% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_sas}
\alias{write_sas}
\title{Create a .zip for SAS (data + sas file)}
\usage{
write_sas(x = NULL, file = NULL)
}
\arguments{
\item{x}{a data.frame}

\item{file}{path to the zip file to save, if missing a .zip
with the same name as the data.frame will be saved in the
current working directory}
}
\description{
Create a .zip for SAS (data + sas file)
}

